refactor: merge `monthViewHelper` and `weekViewHelper` to `viewHelper`
